# doc/api.yaml
openapi: 3.0.1
info:
  title: mcs api
  description: list of main api
  version: 1.0.0
servers:
- url: https://mcs.cloud273.com/app/api
paths:
...
  /clinic:
    get:
      tags:
      - clinic/information
      summary: Get clinic information
      operationId: getClinicByClinic
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicDetailWithCert'   
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
...
  /clinic/doctor:
    get:
      tags:
      - clinic/information
      summary: Get doctor
      operationId: getDoctorByClinic
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      - name: id
        in: query
        description: doctorId
        required: true
        schema:
          type: number
          default: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorDetailWithCert'   
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
  /clinic/doctor/package/list:
    get:
      tags:
      - clinic/information
      summary: Get list of package of a doctor
      operationId: getListPackageOfDoctorByClinic
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      - name: doctorId
        in: query
        description: doctorId
        required: true
        schema:
          type: number
          default: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackageBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found doctor
        500:
          description: Unknown
  /clinic/doctor/working-time/list:
    get:
      tags:
      - clinic/information
      summary: Get list of working time of a doctor
      operationId: getListWorkingTimeOfDoctorByClinic
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      - name: doctorId
        in: query
        description: doctorId
        required: true
        schema:
          type: number
          default: 1
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2019-12-12T07%3A00%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-31T07%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkingTimeDetail'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found doctor
        500:
          description: Unknown
  /clinic/doctor/package:
    get:
      tags:
      - clinic/information
      summary: Get package of doctor (include active schedule)
      operationId: getPackageWithTimeOfDoctorByClinic
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      - name: id
        in: query
        description: packageId
        required: true
        schema:
          type: number
          default: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageDetailWithSchedule'   
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
...
  /doctor:
    get:
      tags:
      - doctor/information
      summary: Get doctor account
      operationId: getDoctorAccount
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorDetailWithCert'   
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /doctor/info:
    get:
      tags:
      - doctor/information
      summary: Get doctor, clinic information
      operationId: getDoctorClinicCountInfo
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  doctor:
                      $ref: '#/components/schemas/DoctorDetailWithCert'
                  clinic:
                      $ref: '#/components/schemas/ClinicDetailWithCert'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /doctor/clinic:
    get:
      tags:
      - doctor/information
      summary: Get clinic information
      operationId: getClinicByDoctor
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicDetailWithCert'   
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /doctor/package/list:
    get:
      tags:
      - doctor/information
      summary: Get list package of doctor
      operationId: getListPackageByDoctor
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackageBasic'   
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /doctor/working-time/list:
    get:
      tags:
      - doctor/information
      summary: Get list working time of doctor
      operationId: getListWorkingTimeByDoctor
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      - name: from
        in: query
        description: from
        required: true
        schema:
          type: string
          default: "2019-12-12T07%3A00%3A00%2B07%3A00"
      - name: to
        in: query
        description: to
        required: true
        schema:
          type: string
          default: "2020-01-31T07%3A00%3A00%2B07%3A00"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkingTimeDetail'   
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
  /doctor/package:
    get:
      tags:
      - doctor/information
      summary: Get package (include active schedule)
      operationId: getPackageByDoctor
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.doctor|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "doctor_token_here"
      - name: id
        in: query
        description: packageId
        required: true
        schema:
          type: integer
          default: 1
          example: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageDetailWithSchedule'   
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
...
  /config:
    get:
      tags:
      - application/config
      summary: Get config
      operationId: getConfig
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  countries:
                    type: array
                    items:
                      $ref: '#/components/schemas/Country'
                  specialties:
                    type: array
                    items:
                      $ref: '#/components/schemas/SpecialtyConfig'
                  reasons:
                    type: object
                    $ref: '#/components/schemas/Reason'
                  appointment:
                    type: object
                    properties:
                      acceptable: 
                        type: number
                        example: -21600
                        default: -21600
                      cancelable: 
                        type: number
                        example: -14400
                        default: -14400
                      rejectable: 
                        type: number
                        example: -14400
                        default: -14400
                      beginable: 
                        type: object
                        properties:
                          from:
                            type: number
                            example: -14400
                            default: -14400
                          to:
                            type: number
                            example: 14400
                            default: 14400
                      finishable: 
                        type: number
                        example: 28800
                        default: 28800
        400:
          description: Bad request
        500:
          description: Unknown
  /config/reasons:
    get:
      tags:
      - application/config
      summary: Get reasons cancel/reject appointment
      operationId: getReasons
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Reason'
        400:
          description: Bad request
        500:
          description: Unknown
  /config/countries:
    get:
      tags:
      - application/config
      summary: Get country config
      operationId: getCountryConfig
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
        400:
          description: Bad request
        500:
          description: Unknown
  /config/specialties:
    get:
      tags:
      - application/config
      summary: Get specialty config
      operationId: getSpecialtyConfig
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpecialtyConfig'
        400:
          description: Bad request
        500:
          description: Unknown
...
components:
  schemas:
    Line:
      maxLength: 512
      minLength: 2
      type: string
      example: "29 Binh Phu P10 Q6"
      default: "29 Binh Phu P10 Q6"
...
    Dob:
      type: string
      format: yyyy-MM-dd
      example: "2000-01-01"
      default: "2000-01-01"
    Bool:
      type: boolean
      example: true
      default: true
    CityCode:
      maxLength: 12
      type: string
      example: "VNSGO001"
      default: "VNSGO001"
    ClinicCertType:
      type: string
      example: working
      default: working
      enum: [working, other]
...
    CountryCode:
      maxLength: 2
      type: string
      example: "VN"
      default:  "VN"
      enum: ["VN"]
    CurrencyUnit:
      type: string
      example: vnd
      default: vnd
      enum: [vnd, usd]
    Date:
      type: string
      format: yyyy-MM-dd
      example: "2020-01-01"
      default: "2020-01-01"
    DateTime:
      type: string
      format: yyyy-MM-ddTHH:mm:ssZ
      example: "2020-01-01T08:00:00+07:00"
      default: "2020-01-01T08:00:00+07:00"
...
    DoctorCertType:
      type: string
      example: working
      default: working
      enum: [personal, working, degree, other]
    Email:
      maxLength: 128
      type: string
      example: "nglequduph@gmail.com"
      default: "nglequduph@gmail.com"
    Firstname:
      maxLength: 128
      minLength: 2
      type: string
      example: "Dung"
      default: "Dung"
    Gender:
      type: string
      example: male
      default: male
      enum: [male, female]
    ID:
      type: integer
      format: int64
      example: 1
      default: 1
    ImageUrl:
      maxLength: 128
      minLength: 4
      type: string
      example: https://cloud273.com/mcs/resource/image.png
      default: https://cloud273.com/mcs/resource/image.png
    Language:
      type: string
      example: vi
      default: vi
      enum: [vi, en]
    Lastname:
      maxLength: 128
      type: string
      example: "Nguyen"
      default: "Nguyen"
    LongLait:
      type: number
      format: float
      example: 10.001
      default: 10.001
...
    Description:
      maxLength: 1000
      type: string
      example: "Description"
      default: "Description"
    Number:
      type: number
      example: 10
      default: 10
    OtherDate:
      type: string
      format: yyyy-MM-dd
      example: "2020-01-31"
      default: "2020-01-31"
...
    PackageType:
      type: string
      example: "classic"
      default: "classic"
      enum: ["classic", "telemed"]
...
    Phone:
      maxLength: 16
      type: string
      example: "+84933270384"
      default: "+84933270384" 
      description: "Phone or mobile"
    SpecialtyCode:
      type: string
      example: "primary"
      default: "primary"
      enum: ["primary", "cardiology"]
    StateCode:
      maxLength: 6
      type: string
      example: "VNSGO"
      default:  "VNSGO"
      enum: ["VNSGO"]
...
    Timezone:
      type: string
      example: "+07:00"
      default: "+07:00"
...
    Username:
      maxLength: 128
      type: string
      example: "nglequduph@gmail.com"
      default: "nglequduph@gmail.com"
      description: "Phone or email"
    VisitTime:
      type: number
      example: 600
      default: 600
...
    Address:
      type: object
      required: [country, state, city, line]
      properties:
        country:
          $ref: '#/components/schemas/CountryCode'
        state:
          $ref: '#/components/schemas/StateCode'
        city:
          $ref: '#/components/schemas/CityCode'
        line:
          $ref: '#/components/schemas/Line'
        longitude:
          $ref: '#/components/schemas/LongLait'
        latitude:
          $ref: '#/components/schemas/LongLait'
    Certificate:
      type: object
      required: [code, name, issuer, issueDate, image]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 32
          example: "100000"
          default: "100000"
        name:
          type: string
          minLength: 6
          maxLength: 64
          example: Certificate of
          default: Certificate of
        issuer:
          type: string
          minLength: 2
          maxLength: 64
          example: SYT HCM
          default: SYT HCM
        issueDate:
          $ref: '#/components/schemas/Date'
        expDate:
          $ref: '#/components/schemas/OtherDate'
        image:
          $ref: '#/components/schemas/ImageUrl'
    DateRange:
      type: object
      required: [from, to]
      properties:
        from:
          $ref: '#/components/schemas/Date'
        to:
          $ref: '#/components/schemas/OtherDate'
    Price:
      type: object
      required: [amount, currency]
      properties:
        amount:
          $ref: '#/components/schemas/Number'
        currency:
          $ref: '#/components/schemas/CurrencyUnit'
    Profile:
      type: object
      required: [firstname, lastname, gender, dob]
      properties:
        firstname:
          $ref: '#/components/schemas/Firstname'
        lastname:
          $ref: '#/components/schemas/Lastname'
        gender:
          $ref: '#/components/schemas/Gender'
        dob:
          $ref: '#/components/schemas/Dob'      
...
    City:
      type: object
      required: [code, name]
      properties:
        code:
          $ref: '#/components/schemas/CityCode'
        name:
          $ref: '#/components/schemas/LocalizeText'
    Country:
      type: object
      required: [code, name, state]
      properties:
        code:
          $ref: '#/components/schemas/CountryCode'
        name:
          $ref: '#/components/schemas/LocalizeText'
        state:
          type: array
          items:
            $ref: '#/components/schemas/State'
    Value:
      type: object
      required: [code, name]
      properties:
        code:
          $ref: '#/components/schemas/PackageType'
        name:
          $ref: '#/components/schemas/LocalizeText'
    Reason:
      type: object
      required: [code, name]
      properties:
        patientCancel:
          type: array
          items:
            $ref: '#/components/schemas/Value'
        clinicReject:
          type: array
          items:
            $ref: '#/components/schemas/Value'
        systemReject:
          type: array
          items:
            $ref: '#/components/schemas/Value'   
    SpecialtyConfig:
      type: object
      required: [code, image, name]
      properties:
        code:
          $ref: '#/components/schemas/SpecialtyCode'
        image:
          $ref: '#/components/schemas/ImageUrl'
        name:
          $ref: '#/components/schemas/LocalizeText'
    State:
      type: object
      required: [code, name, city]
      properties:
        code:
          $ref: '#/components/schemas/StateCode'
        name:
          $ref: '#/components/schemas/LocalizeText'
        city:
          type: array
          items:
            $ref: '#/components/schemas/City'
    LocalizeText:
      type: object
      required: [vi, en]
      properties:
        vi:
          type: string
          example: vietnamese text
          default: vietnamese text 
        en:
          type: string
          example: english text
          default: english text 
...
    ClinicBasic:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
          minLength: 6
          maxLength: 64
          example: Clinic name
          default: Clinic name
    ClinicDetail:
      allOf:
        - $ref: '#/components/schemas/ClinicBasic'
      type: object
      properties:
        email:
          $ref: '#/components/schemas/Email'
        phone:
          $ref: '#/components/schemas/Phone'
        workPhone:
          $ref: '#/components/schemas/Phone'
        address:
          $ref: '#/components/schemas/Address'
        image:
          $ref: '#/components/schemas/ImageUrl'
        description:
          $ref: '#/components/schemas/Description'
        deactivated:
          $ref: '#/components/schemas/Bool'
    ClinicDetailWithCert:
      allOf:
        - $ref: '#/components/schemas/ClinicDetail'
      type: object
      properties:
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/ClinicCertDetail'
...
    ClinicCertUpdateWithID:
      allOf:
        - $ref: '#/components/schemas/Certificate'
      type: object
      required: [type, id]
      properties:
        type:
          $ref: '#/components/schemas/ClinicCertType'
        id:
          $ref: '#/components/schemas/ID'
        description:
          $ref: '#/components/schemas/Description'
    ClinicCertBasic:
      allOf:
        - $ref: '#/components/schemas/ClinicCertUpdateWithID'
    ClinicCertDetail:
      allOf:
        - $ref: '#/components/schemas/ClinicCertBasic'
...
    DoctorBasic:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        image:
          $ref: '#/components/schemas/ImageUrl'
        profile:
          $ref: '#/components/schemas/Profile'
        specialties:
          type: array
          items:
            $ref: '#/components/schemas/SpecialtyCode'
        title:
          type: string
          maxLength: 32
          default: Dr
          example: Dr
        office:
          type: string
          maxLength: 256
          default: CEO Hospital
          example: CEO Hospital
        startWork:
          $ref: '#/components/schemas/Date'
    DoctorDetail:
      allOf:
        - $ref: '#/components/schemas/DoctorBasic'
      type: object
      properties:
        language:
          $ref: '#/components/schemas/Language'
        username:
          $ref: '#/components/schemas/Username'
        specialties:
          type: array
          items:
            $ref: '#/components/schemas/SpecialtyCode'
        biography:
          type: string
          maxLength: 10000
          default: Something about doctor's biography
          example: Something about doctor's biography
        timezone:
          $ref: '#/components/schemas/Timezone'
        description:
          $ref: '#/components/schemas/Description'
        deactivated:
          $ref: '#/components/schemas/Bool'
    DoctorDetailWithCert:
      allOf:
        - $ref: '#/components/schemas/DoctorDetail'
      type: object
      properties:
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/DoctorCertDetail'
...
    DoctorCertUpdateWithID:
      allOf:
        - $ref: '#/components/schemas/Certificate'
      type: object
      required: [type, id]
      properties:
        type:
          $ref: '#/components/schemas/DoctorCertType'
        id:
          $ref: '#/components/schemas/ID'
        description:
          $ref: '#/components/schemas/Description'
    DoctorCertBasic:
      allOf:
        - $ref: '#/components/schemas/DoctorCertUpdateWithID'
    DoctorCertDetail:
      allOf:
        - $ref: '#/components/schemas/DoctorCertBasic'
...
    PackageBasic:
      type: object
      required: [id, specialty, type, price, visitTime]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        specialty:
          $ref: '#/components/schemas/SpecialtyCode'
        type:
          $ref: '#/components/schemas/PackageType'
        price:
          $ref: '#/components/schemas/Price'
        visitTime:
          $ref: '#/components/schemas/VisitTime'
        description:
          $ref: '#/components/schemas/Description'
    PackageDetail:
      allOf:
        - $ref: '#/components/schemas/PackageBasic'
    PackageDetailWithSchedule:
      allOf:
        - $ref: '#/components/schemas/PackageDetail'
      type: object
      properties:
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleDetail'
...
    ScheduleBase:
      type: object
      required: [duration, monday, tuesday, wednesday, thursday, friday, saturday, sunday]
      properties:
        duration:
          $ref: '#/components/schemas/DateRange'
        monday:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TimeRange'
        tuesday:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TimeRange'
        wednesday:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TimeRange'
        thursday:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TimeRange'
        friday:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TimeRange'
        saturday:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TimeRange'
        sunday:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TimeRange'
...
    ScheduleBasic:
      allOf:
        - $ref: '#/components/schemas/ScheduleBase'
      type: object
      required: [id]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        description:
          $ref: '#/components/schemas/Description'
    ScheduleDetail:
      allOf:
        - $ref: '#/components/schemas/ScheduleBasic'
...
    WorkingTimeDetail:
      type: object
      required: [begin, end]
      properties:
        begin:
          $ref: '#/components/schemas/DateTime'
        end:
          $ref: '#/components/schemas/DateTime'
        packageId:
          $ref: '#/components/schemas/Number'
        scheduleId:
          $ref: '#/components/schemas/Number'
        workingDayId:
          $ref: '#/components/schemas/Number'
        visitTime:
          $ref: '#/components/schemas/Number'